+++ interpreter/src/constante.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 use crate::traits::GetResult;
@@ -5,11 +6,11 @@ use crate::traits::GetResult;
 #[derive(Deserialize, Serialize)]
 pub struct Constant {
     name: String,
-    value: f64,
+    value: Decimal,
 }
 
 impl Constant {
-    pub fn new(name: &str, value: f64) -> Self {
+    pub fn new(name: &str, value: Decimal) -> Self {
         Constant {
             name: name.to_string(),
             value,
@@ -17,8 +18,8 @@ impl Constant {
     }
 }
 
-impl GetResult<Option<f64>> for Vec<Constant> {
-    fn get_result(&self, input: &str) -> Option<f64> {
+impl GetResult<Option<Decimal>> for Vec<Constant> {
+    fn get_result(&self, input: &str) -> Option<Decimal> {
         self.iter()
             .find(|history| history.name == input)
             .map(|history| history.value)
